promptdna/backend/requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
pydantic==1.10.7
python-multipart==0.0.5
transformers==4.35.0
sentence-transformers==2.2.2
faiss-cpu==1.8.0
pymongo==4.4.0
dnspython==2.3.0
requests==2.31.0
python-dotenv==1.0.0
openai==1.29.0
